api/requirements.txt
wheel
fastapi
uvicorn
asyncpg
tortoise-orm
environs
httpx
tonsdk